Separated api routes
