+++ README.md
@@ -9,7 +9,7 @@
 ### Both
 
 * Install Node.js
-* `npm install -g bower gulp coffee-script`
+* `npm install -g bower gulp coffee-script pug stylus`
 * `npm install` - node dependencies
 * `bower install` - frontend dependencies
 * `pip install -r requirements.txt` - to install all dependencies
@@ -17,6 +17,7 @@
 ## Database
 
 * Install PostgreSQL
+* Enter shell: 'psql -U postgres'
 * `createdb data-vis`
 * `createuser data-vis -w 'data-vis'`
