+++ helpers/export/import.json
@@ -1,0 +1,76 @@
+{
+    "cities": [
+        {
+            "id": 1,
+            "name": "Київ",
+            "lat": 50.447,
+            "lon": 30.550
+        },
+        {
+            "id": 2,
+            "name": "Львів",
+            "lat": 49.842,
+            "lon": 24.033
+        },
+        {
+            "id": 3,
+            "name": "Рівне",
+            "lat": 50.618,
+            "lon": 26.250
+        }
+    ],
+    "data_sets": [
+        {
+            "id": 1,
+            "name": "Народжуваність"
+        },
+        {
+            "id": 2,
+            "name": "Смертність"
+        }
+    ],
+    "data_entries" : [
+        {
+            "id": 1,
+            "value": 42,
+            "time": 1458421488,
+            "set_id": 1,
+            "city_id": 1
+        },
+        {
+            "id": 2,
+            "value": 44,
+            "time": 1458421488,
+            "set_id": 1,
+            "city_id": 1
+        },
+        {
+            "id": 3,
+            "value": 43,
+            "time": 1458421488,
+            "set_id": 1,
+            "city_id": 1
+        },
+        {
+            "id": 4,
+            "value": 334,
+            "time": 1458421488,
+            "set_id": 2,
+            "city_id": 1
+        },
+        {
+            "id": 5,
+            "value": 456,
+            "time": 1458421488,
+            "set_id": 2,
+            "city_id": 1
+        },
+        {
+            "id": 6,
+            "value": 666,
+            "time": 1458421488,
+            "set_id": 2,
+            "city_id": 1
+        }
+    ]
+}
