+++ helpers/sql/drop_tables.sql
@@ -1,0 +1,3 @@
+DROP TABLE IF EXISTS data_entries;
+DROP TABLE IF EXISTS data_sets;
+DROP TABLE IF EXISTS cities;
