+++ models/data_sets.py
@@ -1,9 +1,6 @@
-from helpers import database
+from helpers import database as db
 
 def get_all():
     sql_query = 'SELECT id, name FROM data_sets'
 
-    cursor = database.conn.cursor()
-    cursor.execute(sql_query)
-    database.conn.commit()
-    return cursor.fetchall()
+    return db.query(sql_query)
